got all but effects working now

+++ code/modules/asset_cache/asset_list_items.dm
@@ -262,15 +262,16 @@
 /datum/asset/spritesheet/vending_products/register()
 	for (var/k in GLOB.vending_products)
 		var/atom/item = k
-		var/icon_file
-		var/icon_state
+		var/icon_file = initial(item.icon)
+		var/icon_state = initial(item.icon_state)
 		var/icon/I
 
 		if (!ispath(item, /atom))
+			world.log << "not atom! [item]"
 			continue
 
-		icon_file = initial(item.icon)
-		icon_state = initial(item.icon_state)
+		if (sprites[icon_file])
+			continue
 
 		if(icon_state in icon_states(icon_file))
 			I = icon(icon_file, icon_state, SOUTH)
@@ -278,10 +279,12 @@
 			if (!isnull(c) && c != "#FFFFFF")
 				I.Blend(initial(c), ICON_MULTIPLY)
 		else
-			item = new item()
+			if(ispath(k, /obj/effect))
+				continue
+			item = new k()
 			I = icon(item.icon, item.icon_state, SOUTH)
 			qdel(item)
-		var/imgid = replacetext(replacetext("[item]", "/obj/item/", ""), "/", "-")
+		var/imgid = replacetext(replacetext("[k]", "/obj/item/", ""), "/", "-")
 
 		Insert(imgid, I)
 	return ..()
